+++ airflow/providers/google/cloud/hooks/bigquery_dts.py
@@ -23,7 +23,7 @@ from typing import Sequence
 
 from google.api_core.gapic_v1.method import DEFAULT, _MethodDefault
 from google.api_core.retry import Retry
-from google.cloud.bigquery_datatransfer_v1 import DataTransferServiceClient
+from google.cloud.bigquery_datatransfer_v1 import DataTransferServiceAsyncClient, DataTransferServiceClient
 from google.cloud.bigquery_datatransfer_v1.types import (
     StartManualTransferRunsResponse,
     TransferConfig,
@@ -32,7 +32,11 @@ from google.cloud.bigquery_datatransfer_v1.types import (
 from googleapiclient.discovery import Resource
 
 from airflow.providers.google.common.consts import CLIENT_INFO
-from airflow.providers.google.common.hooks.base_google import PROVIDE_PROJECT_ID, GoogleBaseHook
+from airflow.providers.google.common.hooks.base_google import (
+    PROVIDE_PROJECT_ID,
+    GoogleBaseAsyncHook,
+    GoogleBaseHook,
+)
 
 
 def get_object_id(obj: dict) -> str:
@@ -263,3 +267,70 @@ class BiqQueryDataTransferServiceHook(GoogleBaseHook):
         return client.get_transfer_run(
             request={"name": name}, retry=retry, timeout=timeout, metadata=metadata or ()
         )
+
+
+class AsyncBiqQueryDataTransferServiceHook(GoogleBaseAsyncHook):
+    """Hook of the BigQuery service to be used with async client of the Google library."""
+
+    sync_hook_class = BiqQueryDataTransferServiceHook
+
+    def __init__(
+        self,
+        gcp_conn_id: str = "google_cloud_default",
+        delegate_to: str | None = None,
+        location: str | None = None,
+        impersonation_chain: str | Sequence[str] | None = None,
+    ):
+        super().__init__(
+            gcp_conn_id=gcp_conn_id,
+            delegate_to=delegate_to,
+            location=location,
+            impersonation_chain=impersonation_chain,
+        )
+        self._conn: DataTransferServiceAsyncClient | None = None
+
+    async def _get_conn(self) -> DataTransferServiceAsyncClient:
+        if not self._conn:
+            credentials = (await self.get_sync_hook()).get_credentials()
+            self._conn = DataTransferServiceAsyncClient(credentials=credentials, client_info=CLIENT_INFO)
+        return self._conn
+
+    async def _get_project_id(self) -> str:
+        sync_hook = await self.get_sync_hook()
+        return sync_hook.project_id
+
+    async def get_transfer_run(
+        self,
+        config_id: str,
+        run_id: str,
+        project_id: str | None,
+        retry: Retry | _MethodDefault = DEFAULT,
+        timeout: float | None = None,
+        metadata: Sequence[tuple[str, str]] = (),
+    ):
+        """
+        Returns information about the particular transfer run.
+
+        :param run_id: ID of the transfer run.
+        :param config_id: ID of transfer config to be used.
+        :param project_id: The BigQuery project id where the transfer configuration should be
+            created. If set to None or missing, the default project_id from the Google Cloud connection
+            is used.
+        :param retry: A retry object used to retry requests. If `None` is
+            specified, requests will not be retried.
+        :param timeout: The amount of time, in seconds, to wait for the request to
+            complete. Note that if retry is specified, the timeout applies to each individual
+            attempt.
+        :param metadata: Additional metadata that is provided to the method.
+        :return: An ``google.cloud.bigquery_datatransfer_v1.types.TransferRun`` instance.
+        """
+        project_id = project_id or (await self._get_project_id())
+        client = await self._get_conn()
+        name = f"projects/{project_id}/transferConfigs/{config_id}/runs/{run_id}"
+        transfer_run = await client.get_transfer_run(
+            name=name,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+        return transfer_run