+++ tests/providers/amazon/aws/hooks/test_datasync.py
@@ -23,7 +23,7 @@ import boto3
 import pytest
 from moto import mock_datasync
 
-from airflow.exceptions import AirflowTaskTimeout
+from airflow.exceptions import AirflowException, AirflowTaskTimeout
 from airflow.providers.amazon.aws.hooks.datasync import DataSyncHook
 
 
@@ -96,6 +96,38 @@ class TestDataSyncHookMocked:
         assert not self.hook.locations
         assert not self.hook.tasks
         assert self.hook.wait_interval_seconds == 0
+
+    @pytest.mark.parametrize(
+        "location_uri, expected_method",
+        [
+            pytest.param("smb://spam/egg/", "create_location_smb", id="smb"),
+            pytest.param("s3://foo/bar", "create_location_s3", id="s3"),
+            pytest.param("nfs://server:2049/path", "create_location_nfs", id="nfs"),
+            pytest.param("efs://12345.efs.aws-region.amazonaws.com/path", "create_location_efs", id="efs"),
+        ],
+    )
+    def test_create_location_method_mapping(self, mock_get_conn, location_uri, expected_method):
+        """Test expected location URI and mapping with DataSync.Client methods."""
+        mock_get_conn.return_value = self.client
+        assert hasattr(self.client, expected_method), f"{self.client} doesn't have method {expected_method}"
+        with mock.patch.object(self.client, expected_method) as m:
+            self.hook.create_location(location_uri, foo="bar", spam="egg")
+            m.assert_called_once_with(foo="bar", spam="egg")
+
+    @pytest.mark.parametrize(
+        "location_uri",
+        [
+            pytest.param("hdfs://namenodehost1/path", id="hdfs"),
+            pytest.param("https://example.org/path", id="https"),
+            pytest.param("http://example.org/path", id="http"),
+            pytest.param("lustre://mount/path", id="lustre"),
+        ],
+    )
+    def test_create_location_unknown_type(self, mock_get_conn, location_uri):
+        """Test unsupported location URI."""
+        mock_get_conn.return_value = mock.MagicMock()
+        with pytest.raises(AirflowException, match="Invalid/Unsupported location type: .*"):
+            self.hook.create_location(location_uri, foo="bar", spam="egg")
 
     def test_create_location_smb(self, mock_get_conn):
         # ### Configure mock: